foundations trbfv refactoring draft
