+++ crates/fhe/src/trbfv/mod.rs
@@ -1,12 +1,21 @@
 //! The Threshold BFV scheme, as described by Antoine Urban and Matthieu Rambaud.
 //! in [Robust Multiparty Computation from Threshold Encryption Based on RLWE](https://eprint.iacr.org/2024/1285.pdf).
 
+/// Configuration and validation for threshold BFV
+pub mod config;
 /// Error types for threshold operations
 pub mod errors;
-/// Traits for threshold secret sharing operations
-pub mod traits;
-/// Main threshold BFV implementation
-pub mod trbfv;
+/// Secret sharing abstractions and implementations
+pub mod secret_sharing;
+/// Share collection and management
+pub mod shares;
+/// Smudging abstractions and implementations
+pub mod smudging;
+/// Main threshold BFV orchestrator
+pub mod threshold;
 
-pub use traits::*;
-pub use trbfv::TRBFV;
+// Re-export main types for convenience
+pub use secret_sharing::{SecretSharer, ShamirSecretSharing};
+pub use shares::ShareManager;
+pub use smudging::{SmudgingConfig, SmudgingGenerator, StandardSmudgingGenerator};
+pub use threshold::TRBFV;
